@@ -4,23 +4,28 @@ apply plugin: 'com.vanniktech.android.junit.jacoco'
 
 buildscript {
   ext.versions = [
-    supportLibrary: '27.0.1',
-    espresso: '3.0.1',
+    espresso: '3.1.1',
+    testRules: '1.1.1',
   ]
 
   ext.deps = [
     'support': [
-      'annotations': "com.android.support:support-annotations:${versions.supportLibrary}",
-      'appcompat': "com.android.support:appcompat-v7:${versions.supportLibrary}",
+      'annotations': 'androidx.annotation:annotation:1.0.1',
+      'appCompat': 'androidx.appcompat:appcompat:1.0.0',
 
       'test': [
-        'espresso': "com.android.support.test.espresso:espresso-core:${versions.espresso}",
+        'espresso': [
+          'core': "androidx.test.espresso:espresso-core:${versions.espresso}",
+        ],
+        'junitExt': "androidx.test.ext:junit:1.1.0",
+        'runner': "androidx.test:runner:${versions.testRules}",
+        'rules': "androidx.test:rules:${versions.testRules}",
       ],
     ],
     'test': [
       'privateconstructor': 'com.pushtorefresh.java-private-constructor-checker:checker:1.2.0',
       'junit': 'junit:junit:4.12',
-      'assertj3': 'org.assertj:assertj-core:3.8.0',
+      'assertj3': 'org.assertj:assertj-core:3.11.1',
     ],
   ]
 
@@ -31,12 +36,12 @@ buildscript {
   }
 
   dependencies {
-    classpath 'com.vanniktech:gradle-code-quality-tools-plugin:0.9.0'
-    classpath 'com.vanniktech:gradle-android-junit-jacoco-plugin:0.11.0'
+    classpath 'com.vanniktech:gradle-code-quality-tools-plugin:0.15.0'
+    classpath 'com.vanniktech:gradle-android-junit-jacoco-plugin:0.13.0'
 
-    classpath 'com.android.tools.build:gradle:3.0.1'
-    classpath 'com.github.ben-manes:gradle-versions-plugin:0.17.0'
-    classpath 'com.vanniktech:gradle-maven-publish-plugin:0.1.0'
+    classpath 'com.android.tools.build:gradle:3.2.1'
+    classpath 'com.github.ben-manes:gradle-versions-plugin:0.20.0'
+    classpath 'com.vanniktech:gradle-maven-publish-plugin:0.6.0'
   }
 }
 
@@ -59,19 +64,20 @@ subprojects {
   repositories {
     mavenCentral()
     google()
+    jcenter()
   }
 }
 
 ext {
   minSdkVersion = 15
-  compileSdkVersion = 27
-  targetSdkVersion = 27
+  compileSdkVersion = 28
+  targetSdkVersion = 28
 
   versionCode = VERSION_CODE
   versionName = VERSION_NAME
 }
 
-task wrapper(type: Wrapper) {
-  gradleVersion = '4.8'
+wrapper {
+  gradleVersion = '4.10.3'
   distributionType = Wrapper.DistributionType.ALL
 }
