@@ -1,8 +1,8 @@
 package com.vanniktech.espresso.core.utils;
 
-import android.support.annotation.DrawableRes;
-import android.support.test.rule.ActivityTestRule;
-import android.support.test.runner.AndroidJUnit4;
+import androidx.annotation.DrawableRes;
+import androidx.test.ext.junit.runners.AndroidJUnit4;
+import androidx.test.rule.ActivityTestRule;
 import com.vanniktech.espresso.core.utils.test.R;
 import junit.framework.AssertionFailedError;
 import org.junit.Rule;
@@ -10,10 +10,10 @@ import org.junit.Test;
 import org.junit.rules.ExpectedException;
 import org.junit.runner.RunWith;
 
-import static android.support.test.espresso.Espresso.onView;
-import static android.support.test.espresso.assertion.ViewAssertions.matches;
-import static android.support.test.espresso.matcher.ViewMatchers.withId;
 import static android.view.View.GONE;
+import static androidx.test.espresso.Espresso.onView;
+import static androidx.test.espresso.assertion.ViewAssertions.matches;
+import static androidx.test.espresso.matcher.ViewMatchers.withId;
 import static com.vanniktech.espresso.core.utils.DrawableMatcher.withDrawable;
 import static com.vanniktech.espresso.core.utils.DrawableMatcher.withNoDrawable;
 import static com.vanniktech.espresso.core.utils.DrawableMatcherActivity.VIEW_ID;
