@@ -1,7 +1,7 @@
 package com.vanniktech.espresso.core.utils;
 
-import android.support.test.rule.ActivityTestRule;
-import android.support.test.runner.AndroidJUnit4;
+import androidx.test.ext.junit.runners.AndroidJUnit4;
+import androidx.test.rule.ActivityTestRule;
 import com.vanniktech.espresso.core.utils.test.R;
 import junit.framework.AssertionFailedError;
 import org.junit.Rule;
@@ -11,9 +11,9 @@ import org.junit.runner.RunWith;
 
 import static android.graphics.Color.GREEN;
 import static android.graphics.Color.RED;
-import static android.support.test.espresso.Espresso.onView;
-import static android.support.test.espresso.assertion.ViewAssertions.matches;
-import static android.support.test.espresso.matcher.ViewMatchers.withId;
+import static androidx.test.espresso.Espresso.onView;
+import static androidx.test.espresso.assertion.ViewAssertions.matches;
+import static androidx.test.espresso.matcher.ViewMatchers.withId;
 import static com.vanniktech.espresso.core.utils.HintTextColorMatcher.withHintTextColor;
 import static com.vanniktech.espresso.core.utils.HintTextColorMatcher.withHintTextColorRes;
 import static com.vanniktech.espresso.core.utils.HintTextColorMatcherActivity.VIEW_ID;
