@@ -1,18 +1,18 @@
 package com.vanniktech.espresso.core.utils;
 
-import android.support.test.rule.ActivityTestRule;
-import android.support.test.runner.AndroidJUnit4;
+import androidx.test.ext.junit.runners.AndroidJUnit4;
+import androidx.test.rule.ActivityTestRule;
 import junit.framework.AssertionFailedError;
 import org.junit.Rule;
 import org.junit.Test;
 import org.junit.rules.ExpectedException;
 import org.junit.runner.RunWith;
 
-import static android.support.test.espresso.Espresso.onView;
-import static android.support.test.espresso.assertion.ViewAssertions.matches;
-import static android.support.test.espresso.matcher.ViewMatchers.withId;
-import static com.vanniktech.espresso.core.utils.ProgressMatcher.withProgress;
+import static androidx.test.espresso.Espresso.onView;
+import static androidx.test.espresso.assertion.ViewAssertions.matches;
+import static androidx.test.espresso.matcher.ViewMatchers.withId;
 import static com.vanniktech.espresso.core.utils.ProgressBarActivity.VIEW_ID;
+import static com.vanniktech.espresso.core.utils.ProgressMatcher.withProgress;
 
 @RunWith(AndroidJUnit4.class) public final class ProgressMatcherTest {
   @Rule public final ExpectedException expectedException = ExpectedException.none();
