@@ -2,7 +2,7 @@ package com.vanniktech.espresso.core.utils;
 
 import android.app.Activity;
 import android.os.Bundle;
-import android.support.annotation.Nullable;
+import androidx.annotation.Nullable;
 import android.widget.TextView;
 
 public final class TextViewDrawableMatcherActivity extends Activity {
