@@ -1,10 +1,10 @@
 package com.vanniktech.espresso.core.utils;
 
 import android.graphics.Color;
-import android.support.annotation.CheckResult;
-import android.support.annotation.ColorInt;
-import android.support.annotation.ColorRes;
-import android.support.test.espresso.matcher.BoundedMatcher;
+import androidx.annotation.CheckResult;
+import androidx.annotation.ColorInt;
+import androidx.annotation.ColorRes;
+import androidx.test.espresso.matcher.BoundedMatcher;
 import android.view.View;
 import android.widget.TextView;
 import org.hamcrest.Description;
