@@ -3,9 +3,9 @@ package com.vanniktech.espresso.core.utils;
 import android.annotation.SuppressLint;
 import android.annotation.TargetApi;
 import android.graphics.drawable.Drawable;
-import android.support.annotation.CheckResult;
-import android.support.annotation.DrawableRes;
-import android.support.test.espresso.matcher.BoundedMatcher;
+import androidx.annotation.CheckResult;
+import androidx.annotation.DrawableRes;
+import androidx.test.espresso.matcher.BoundedMatcher;
 import android.view.View;
 import android.widget.TextView;
 import org.hamcrest.Description;
