@@ -1,9 +1,9 @@
 package com.vanniktech.espresso.core.utils;
 
-import android.support.annotation.CheckResult;
-import android.support.test.espresso.matcher.BoundedMatcher;
-import android.support.v4.view.ViewPager;
 import android.view.View;
+import androidx.annotation.CheckResult;
+import androidx.test.espresso.matcher.BoundedMatcher;
+import androidx.viewpager.widget.ViewPager;
 import org.hamcrest.Description;
 
 public final class CurrentItemMatcher extends BoundedMatcher<View, ViewPager> {
